Rename test-data to crypto-test-data since that is what it contains
