+++ spec/integ/crypto/state-events.spec.ts
@@ -21,7 +21,7 @@ import Olm from "@matrix-org/olm";
 
 import * as testUtils from "../../test-utils/test-utils";
 import { getSyncResponse, syncPromise } from "../../test-utils/test-utils";
-import { TEST_ROOM_ID as ROOM_ID } from "../../test-utils/test-data";
+import { TEST_ROOM_ID as ROOM_ID } from "../../test-utils/crypto-test-data";
 import { logger } from "../../../src/logger";
 import {
     createClient,
