+++ spec/integ/crypto/to-device-messages.spec.ts
@@ -28,7 +28,7 @@ import {
     type MatrixEvent,
     type ReceivedToDeviceMessage,
 } from "../../../src";
-import * as testData from "../../test-utils/test-data";
+import * as testData from "../../test-utils/crypto-test-data";
 import { E2EKeyResponder } from "../../test-utils/E2EKeyResponder";
 import { SyncResponder } from "../../test-utils/SyncResponder";
 import { E2EKeyReceiver } from "../../test-utils/E2EKeyReceiver";
